refactor: move rag out of backends
