+++ gerd/qa/qa_service.py
@@ -17,8 +17,8 @@ from langchain.text_splitter import RecursiveCharacterTextSplitter
 from langchain_community.document_loaders import PyPDFLoader, TextLoader
 
 import gerd.loader as gerd_loader
-from gerd.backends.rag import FAISS, Rag, create_faiss, load_faiss
 from gerd.models.qa import QAConfig
+from gerd.rag import FAISS, Rag, create_faiss, load_faiss
 from gerd.transport import (
     DocumentSource,
     FileTypes,
